Add dask.tokenize for sparray classes, move to conftest
